ls-hw-08 update

+++ src/js/03-feedback.js
@@ -1,31 +1,33 @@
 import throttle from 'lodash.throttle';
 
 const form = document.querySelector('.feedback-form');
+const { email, message } = form.elements;
+
+const submitForm = (event) => {
+  event.preventDefault();
+  if (email.value && message.value) {
+    console.log({ email: email.value, message: message.value });
+
+    localStorage.clear();
+
+    email.value = "";
+    message.value = "";
+  }
+}
 
 const storeForm = throttle(() => {
-    localStorage.setItem("feedback-form-state", JSON.stringify({ email: form.elements.email.value, message: form.elements.message.value }));
+    localStorage.setItem("feedback-form-state", JSON.stringify({ email: email.value, message: message.value }));
   }, 500)
 
-form.elements.email.addEventListener('input', storeForm);
+form.addEventListener('input', storeForm);
 
-form.elements.message.addEventListener('input', storeForm);
-
-form.addEventListener('submit', (event) => {
-  event.preventDefault();
-  console.log({ email: form.elements.email.value, message: form.elements.message.value });
-
-  localStorage.clear();
-
-  form.elements.email.value = "";
-  form.elements.message.value = "";
-});
-
+form.addEventListener('submit', submitForm);
 
 const formValue = localStorage.getItem("feedback-form-state");
 
 if (formValue) {
   const formData = JSON.parse(formValue);
-  form.elements.email.value = formData.email;
-  form.elements.message.value = formData.message;
+  email.value = formData.email;
+  message.value = formData.message;
 }
